+++ lyzr/data_analyzr/analysis_handler/plotter.py
@@ -24,17 +24,18 @@ from lyzr.base.llm import LiteLLM
 from lyzr.base.prompt import LyzrPromptFactory
 from lyzr.base.base import UserMessage, SystemMessage
 from lyzr.data_analyzr.models import FactoryBaseClass
+from lyzr.data_analyzr.utils import deterministic_uuid
 from lyzr.data_analyzr.analysis_handler.utils import (
     handle_plotpath,
-    extract_python_code,
     extract_df_names,
+    iterate_llm_calls,
     make_locals_string,
+    extract_python_code,
     extract_column_names,
     remove_print_and_plt_show,
 )
 from lyzr.data_analyzr.db_connector import DatabaseConnector
 from lyzr.data_analyzr.vector_store_utils import ChromaDBVectorStore
-from lyzr.data_analyzr.utils import iterate_llm_calls, deterministic_uuid
 
 default_plot_path = "generated_plots/plot.png"
