+++ lyzr/data_analyzr/utils.py
@@ -15,11 +15,6 @@ from functools import wraps
 import numpy as np
 import pandas as pd
 
-# local imports
-from lyzr.data_analyzr.models import ContextDict
-from lyzr.base.errors import AnalysisFailedError
-from lyzr.base import SystemMessage, AssistantMessage, LiteLLM
-
 
 def deterministic_uuid(content: Union[str, bytes, list] = None):
     """
@@ -237,216 +232,3 @@ def logging_decorator(logger: logging.Logger):
         return wrapper
 
     return decorator_wrapper
-
-
-def iterate_llm_calls(
-    max_retries: int,
-    *,
-    llm: LiteLLM,
-    llm_messages: list,
-    logger: logging.Logger,
-    log_messages: dict = {},
-    time_limit: int = None,
-    llm_kwargs: dict = {},
-):
-    """
-    Decorator to iterate over LLM calls with retry logic and logging.
-
-    This decorator is designed to wrap a function that processes responses from an LLM.
-    It uses the `repeater` function to retry the LLM calls until a valid result is obtained
-    or the maximum number of retries or time limit is reached.
-    The decorator logs the start and end of the process, as well as the input arguments and the result.
-
-    Args:
-        max_retries (int): Maximum number of tries for the LLM call.
-        llm (LiteLLM): Instance of the LLM to be used.
-        llm_messages (list): List of messages to be sent to the LLM.
-        logger (logging.Logger): Logger instance for logging information, errors, and debug messages.
-        log_messages (dict, optional): Custom log messages for start and end of the process. Defaults to {}.
-        time_limit (int, optional): Time limit in seconds for the entire process. Defaults to None.
-        llm_kwargs (dict, optional): Additional keyword arguments to be passed to the LLM. Defaults to {}.
-
-    Returns:
-        function: Wrapped function with retry logic and logging.
-
-    Example:
-        def process_llm_response(llm_response, **kwargs):
-            # Function implementation
-            pass
-        process_llm_response = iterate_llm_calls(
-            max_retries=3,
-            llm=my_llm_instance,
-            llm_messages=my_messages,
-            logger=my_logger,
-            log_messages={"start": "Beginning LLM processing.", "end": "LLM processing finished."},
-            time_limit=60,
-            llm_kwargs={"temperature": 0.7}
-        )(process_llm_response)
-    """
-
-    def decorator_wrapper(func):
-        def wrapped_func(**kwargs):
-            result = None
-            start_time = time.time()
-            logger.info(
-                log_messages.get("start", "Starting LLM analysis."),
-                extra={
-                    "function": func.__name__,
-                    "input_kwargs": {
-                        "max_retries": max_retries,
-                        "start_time": start_time,
-                        "time_limit": time_limit,
-                        "llm": llm,
-                        "llm_messages": llm_messages,
-                        "llm_kwargs": llm_kwargs,
-                        "kwargs": kwargs,
-                    },
-                },
-            )
-            result = repeater(
-                max_retries=max_retries,
-                start_time=start_time,
-                time_limit=time_limit,
-                logger=logger,
-                llm=llm,
-                llm_messages=llm_messages,
-                func=func,
-                llm_kwargs=llm_kwargs,
-                **kwargs,
-            )
-            if result is None:
-                logger.error(
-                    f"Result for {func.__name__} is None",
-                    extra={
-                        "function": func.__name__,
-                        "input_kwargs": kwargs,
-                        "response": result,
-                    },
-                )
-            logger.info(log_messages.get("end", "LLM analysis completed."))
-            return result
-
-        return wrapped_func
-
-    return decorator_wrapper
-
-
-def repeater(
-    max_retries: int,
-    start_time: float,
-    time_limit: int,
-    logger: logging.Logger,
-    llm: LiteLLM,
-    llm_messages: list,
-    func: callable,
-    llm_kwargs: dict = None,
-    **kwargs,
-):
-    """
-    Retries the execution of a function that processes responses from an LLM until a valid
-    result is obtained or the maximum number of retries or time limit is reached.
-
-    Parameters:
-        max_retries (int): The maximum number of retries allowed.
-        start_time (float): The start time of the execution.
-        time_limit (int): The maximum time allowed for execution in seconds.
-        logger (logging.Logger): The logger instance for logging information and errors.
-        llm (LiteLLM): The llm used to generate responses.
-        llm_messages (list): The list of messages to be sent to the llm.
-        func (callable): The function to process the llm response.
-        llm_kwargs (dict, optional): Additional keyword arguments for the LLM. Defaults to None.
-        **kwargs: Additional keyword arguments for the function `func`.
-
-    Returns:
-        Any: The result of the function `func` if successful, otherwise None.
-
-    Raises:
-        AnalysisFailedError: If the execution exceeds the time limit.
-
-    Procedure:
-    - Calls the LLM to generate a response.
-    - Calls the input callable to process the LLM response, and captures the result.
-    - If the result is None, logs an error and retries the process.
-    - If the result is not None, logs the result and returns it.
-    - If there is an error in processing the response
-        - logs the error
-        - appends the LLM response to the messages
-        - appends a system message with the error to the messages
-        - retries the process
-    - If the time limit is exceeded and result remains None, raises an AnalysisFailedError.
-    - If the maximum number of retries is reached and result remains None, returns None.
-    - If the result is obtained, logs the result and returns it.
-    """
-    result = None
-    llm_kwargs = {} if llm_kwargs is None else llm_kwargs
-    max_retries = 1 if max_retries is None else max_retries
-    time_limit = 0 if time_limit is None else time_limit
-    for i in range(max_retries):
-        llm_response = llm.run(messages=llm_messages, **llm_kwargs)
-        if llm_response is None:
-            continue
-        try:
-            if "llm_response" in kwargs:
-                del kwargs["llm_response"]
-            result = func(llm_response=llm_response.message.content, **kwargs)
-            if result is None:
-                logger.error(
-                    f"Result is None in iteration number {i + 1} of function {func.__name__}."
-                )
-                continue
-            logger.info(
-                f"Result from {func.__name__} recieved: {result}",
-                extra={
-                    "function": func.__name__,
-                    "input_kwargs": {
-                        "llm_response": llm_response.message.content,
-                        **kwargs,
-                    },
-                    "response": {
-                        "llm_response": llm_response.message.content,
-                        "result": result,
-                    },
-                },
-            )
-            break
-        except Exception as e:
-            logger.error(
-                f"Error in iteration number {i + 1} of function {func.__name__}. {e.__class__.__name__}: {e}.",
-                extra={
-                    "function": func.__name__,
-                    "input_kwargs": {
-                        "llm_response": llm_response.message.content,
-                        **kwargs,
-                    },
-                    "response": result,
-                    "traceback": traceback.format_exc().splitlines(),
-                },
-            )
-            llm_messages.append(AssistantMessage(content=llm_response.message.content))
-            llm_messages.append(
-                SystemMessage(
-                    content=f"Your response resulted in the following error:\n{traceback.format_exc()}\n\n"
-                    "Read and understand the error CAREFULLY.\nUNDERSTAND how to FIX the error.\nCHANGE your code accordingly."
-                )
-            )
-        finally:
-            if (
-                result is None
-                and time.time() - start_time > time_limit
-                and time_limit > 0
-            ):
-                logger.error(
-                    f"Timeout in iteration number {i + 1} of function {func.__name__} with time limit of {time_limit} seconds.",
-                    extra={
-                        "function": func.__name__,
-                        "input_kwargs": {
-                            "llm_response": llm_response.message.content,
-                            **kwargs,
-                        },
-                        "response": result,
-                    },
-                )
-                raise AnalysisFailedError(
-                    f"Timeout in iteration number {i + 1} of function {func.__name__} with time limit of {time_limit} seconds."
-                )
-    return result
